fix(central-daemon): epoll error check, intent_count type, and callback path condition
- main.c: check epoll_ctl(EPOLL_CTL_ADD) return on new client connection;
  on failure, clean up ctx and close fd instead of silently continuing
- trie.h: change intent_count from int to int64_t to prevent overflow on
  systems with very high concurrent lock counts
- trie.c: remove spurious path_len > 0 condition from parent-grant callback
  in cleanup_fd; also invoke callback with "<unknown>" fallback when the
  released node's own path is unavailable

+++ central-daemon/include/trie.h
@@ -23,7 +23,7 @@ typedef struct trie_node {
 
     /* Lock state */
     int owner_fd;               /* FD of the connection holding an exclusive lock on this path */
-    int intent_count;           /* Number of exclusive locks in the subtree below this node */
+    int64_t intent_count;       /* Number of exclusive locks in the subtree below this node */
     
     /* Wait queue - list of FDs waiting for this node to become available */
     int *waiters;
